Closeable B-Spline!
The B-Spline is not set up that when you try to click the origin point with 4 or more points in the curve, it will automatically set the next point to the origin point and end the command.

Also:
Added some functions to Item to allow all drawable items to flag themselves as showing or hiding points. Flag is called pointVisibility and has two setters and a getter